Removed video section and renamed it to GeneralIntroduction (Added: 3-step icons and descriptions (Still Missing: proper icons from template, reworded descriptions))
