+++ src/components/generalIntroduction/GeneralIntroduction.scss
@@ -1,0 +1,216 @@
+//Variables
+$brand-secondary: #798eea !default;
+
+//== Media queries breakpoints
+//
+//## Define the breakpoints at which your layout will change, adapting to different screen sizes.
+/* ---------------Screen Breakpoints Start ---------------- */
+// Small screen / tablet
+//** Deprecated `$screen-sm` as of v3.0.1
+$screen-sm:                  768px !default;
+$screen-sm-min:              $screen-sm !default;
+/* ---------------Screen Breakpoints Ends ---------------- */
+
+/* ---------------Style Starts--------------- */
+body {
+  font-family: "Poppins", Arial, sans-serif;
+  font-weight: 300;
+  font-size: 16px;
+  line-height: 2;
+  color: gray;
+  background: whitesmoke; }
+ 
+h1, h2, h3, h4, h5, h6, figure {
+  color: #303133;
+  font-family: "Nunito", Arial, sans-serif;
+  font-weight: 400;
+  margin: 0 0 20px 0;
+  line-height: 1.3; }
+
+.colorlib-intro {
+    padding: 7em 0;
+    clear: both;
+    margin-bottom: 2em; 
+}
+.colorlib-intro .colorlib-heading {
+    margin-bottom: 2em; }
+  
+  .colorlib-intro .play {
+    width: 100%;
+    display: block;
+    margin: 30px 0; }
+    .colorlib-intro .play a {
+      width: 100px;
+      height: 100px;
+      display: table;
+      margin: 0 auto;
+      text-align: center;
+      background: #798eea;
+      -webkit-border-radius: 50%;
+      -moz-border-radius: 50%;
+      -ms-border-radius: 50%;
+      border-radius: 50%; }
+      .colorlib-intro .play a .play-button {
+        font-size: 20px;
+        color: #fff;
+        display: table-cell;
+        vertical-align: middle; 
+        padding-top: 35px}
+
+.colorlib-heading {
+    margin-bottom: 5em; }
+    .colorlib-heading h2 {
+        font-size: 50px;
+        font-weight: 300;
+        margin-bottom: 30px;
+        position: relative; }
+        @media screen and (max-width: 768px) {
+        .colorlib-heading h2 {
+            font-size: 30px; } }
+.colorlib-services {
+  padding: 7em 0;
+  clear: both; }
+.colorlib-bg-white {
+  border: 1px solid #e6e6e6;
+  border-left: none;
+  border-right: none;
+  border-top: none;
+  border-bottom: none; }
+
+/* ---------------Style Ends--------------- */
+
+/* ---------------Service Starts--------------*/
+//SERVICES
+.services{
+	margin-bottom: 50px;
+	position: relative;
+	z-index: 0;
+	padding: 0 1em;
+	@media screen and(max-width: $screen-sm){
+		margin-bottom: 3em;
+	}
+	.icon{
+		position: relative;
+		margin-bottom: 30px;
+		display: inline-block;
+		i{
+			font-size: 50px;
+			color: $brand-secondary;
+			z-index: 1;
+		}
+	}
+	.desc{
+		h3{
+			font-size: 26px;
+			font-weight: 300;
+		}
+	}
+}
+/* ---------------Service Ends--------------*/
+
+/* ---------------BootStrap Starts--------------- */
+.container {
+    margin-right: auto;
+    margin-left: auto;
+    padding-left: 15px;
+    padding-right: 15px;
+    max-width: 100%;
+    border-radius: 6px;
+  }
+  .container:before, .container:after {
+    content: " ";
+    display: table;
+  }
+  .container:after {
+    clear: both;
+  }
+  @media (min-width: 768px) {
+    .container {
+      width: 750px;
+      padding-left: 60px;
+      padding-right: 60px;
+    }
+  }
+  @media (min-width: 992px) {
+    .container {
+      width: 970px;
+    }
+  }
+  @media (min-width: 1200px) {
+    .container {
+      width: 1170px;
+    }
+  }
+  
+  .container-fluid {
+    margin-right: auto;
+    margin-left: auto;
+    padding-left: 15px;
+    padding-right: 15px;
+  }
+  .container-fluid:before, .container-fluid:after {
+    content: " ";
+    display: table;
+  }
+  .container-fluid:after {
+    clear: both;
+  }
+
+  .row {
+    margin-left: -15px;
+    margin-right: -15px;
+  }
+  .row:before, .row:after {
+    content: " ";
+    display: table;
+  }
+  .row:after {
+    clear: both;
+  }
+.col-md-8, .col-md-12 {
+    position: relative;
+    min-height: 1px;
+    padding-left: 15px;
+    padding-right: 15px;
+}
+@media (min-width: 992px) {
+  .col-md-12, .col-md-8, .col-md-4{
+    float: left;
+  }
+  .col-md-4 {
+    width: 33.33333%;
+  }
+  .col-md-12 {
+    width: 100%;
+  }
+  .col-md-8 {
+    width: 66.66667%;
+  }
+  .col-md-offset-2 {
+    margin-left: 16.66667%;
+}
+}
+
+.text-center {
+    text-align: center;
+}
+/* ---------------BootStrap Ends--------------- */
+
+/* --------------Glyphicon Starts-----------------*/
+.glyphicon {
+  position: relative;
+  top: 1px;
+  display: inline-block;
+  font-family: 'Glyphicons Halflings';
+  font-style: normal;
+  font-weight: normal;
+  line-height: 1;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+  font-size: 45px;
+  color: $brand-secondary;
+}
+.glyphicon-list-alt               { &:before { content: "\e032"; } }
+.glyphicon-user                   { &:before { content: "\e008"; } }
+.glyphicon-cloud-download         { &:before { content: "\e197"; } }
+/* --------------Glyphicon Ends-----------------*/
